refactor: move rewrite file list
