+++ .changeset/shiny-weeks-clean.md
@@ -1,0 +1,5 @@
+---
+'@sky-fly/template': patch
+---
+
+refactor: move rewrite file list
